# [libc++] Optimize map::insert_or_assign